Moved create_transport() helper to reusable utils file.
